+++ crates/base/src/mob.rs
@@ -30,7 +30,7 @@ pub struct BaseMob {
     pub eva: i16,
     pub fs: f32,
     pub level: i16,
-    pub max_hp: i32,
+    pub max_hp: i16,
     pub max_mp: i16,
     pub mob_type: i16,
     pub pushed: i8,
